backend/requirements.txt
fastapi
uvicorn[standard]
gunicorn
python-multipart
google-genai
opencv-python-headless
pillow
python-dotenv
firebase-admin
pydantic
